visualizing GPS and Temperature data

+++ app/src/main/java/com/logismart/logismart/OnMyChangeListener.java
@@ -1,5 +1,5 @@
 package com.logismart.logismart;
 
 public interface OnMyChangeListener {
-    void onChange(int value);
+    void onChange(float value);
 }
